Completed "send to app" UI and functions
This is the option to send to an external application when conversion
is complete.

note: I cannot complete testing as the "add files to queue" throws an
error for both this fork as well as main project. I have individually
tested each unit, but an end-to-end should be conducted before trusting.